Move ResponsiveFacilityMapTable file
