+++ src/applications/gi/components/profile/ResponsiveFacilityMapTable.jsx
